+++ crates/runtime/src/dialogue.rs
@@ -1,3 +1,5 @@
+//! Adapted from <https://github.com/YarnSpinnerTool/YarnSpinner/blob/da39c7195107d8211f21c263e4084f773b84eaff/YarnSpinner/Dialogue.cs>
+
 use crate::prelude::*;
 use log::error;
 use std::fmt::Debug;
