+++ crates/runtime/src/lib.rs
@@ -1,7 +1,4 @@
 extern crate alloc;
-extern crate core;
-extern crate core;
-extern crate core;
 
 mod dialogue;
 mod dialogue_option;
