+++ crates/runtime/src/markup.rs
@@ -2,5 +2,5 @@ mod attribute_marker_processor;
 mod line_parser;
 mod markup_parse_result;
 
-pub(crate) use self::attribute_marker_processor::*;
-pub use self::{line_parser::*, markup_parse_result::*};
+pub use self::markup_parse_result::*;
+pub(crate) use self::{attribute_marker_processor::*, line_parser::*};
