+++ crates/runtime/src/markup/attribute_marker_processor/dialogue_attribute_marker_processor.rs
@@ -1,0 +1,1 @@
+//! Adapted from <https://github.com/YarnSpinnerTool/YarnSpinner/blob/da39c7195107d8211f21c263e4084f773b84eaff/YarnSpinner/YarnSpinner/Dialogue.cs>, which was split into multiple files.
