+++ crates/runtime/src/markup/markup_parse_result.rs
@@ -104,10 +104,11 @@ impl MarkupParseResult {
                         // attribute begins.
                         let original_length = attribute.length;
                         attribute.length = deletion_start.saturating_sub(start);
-                        if original_length >= 0 && attribute.length <= 0 {
-                            // This attribute is now zero-length. It
-                            // doesn't apply to any text, so we can
-                            // discard it.
+                        if original_length > 0 && attribute.length <= 0 {
+                            // The attribute's length has been reduced to
+                            // zero. All of the contents it previous had
+                            // have been removed, so we will remove the
+                            // attribute itself.
                             return None;
                         }
                     } else {
