+++ crates/yarn_slinger/src/lib.rs
@@ -4,7 +4,7 @@ pub mod prelude {
     pub use yarn_slinger_compiler::prelude::*;
     pub use yarn_slinger_core::generated::*;
     pub use yarn_slinger_core::prelude::{
-        InvalidCastError, Library, Position, Program, Type, YarnFn, YarnFnRegistry,
+        InvalidCastError, Library, Program, Type, YarnFn, YarnFnRegistry,
     };
     pub use yarn_slinger_runtime::prelude::*;
 }
